requirements.txt
# Seal Detection Dependencies (GPU version - CUDA 12.x)

# Core
numpy>=1.21.0
opencv-python>=4.6.0
Pillow>=9.0.0

# PyTorch ecosystem (already in base image)
# torch>=2.5.0
# torchvision>=0.20.0

# YOLOX dependencies
cython
pycocotools>=2.0.6
onnx>=1.15.0
onnxruntime-gpu>=1.17.0
onnxsim>=0.4.0

# Training utilities (YOLOX)
tensorboard>=2.15.0
thop>=0.1.1
albumentations>=1.3.0
scikit-learn>=1.3.0

# Utilities
loguru>=0.7.0
tqdm>=4.66.0
tabulate>=0.9.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Video processing
ffmpeg-python>=0.2.0

# Jupyter support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Config
PyYAML>=6.0

# Scientific
scipy>=1.11.0
pandas>=2.1.0
